+++ mobile/android/android-components/components/feature/prompts/src/main/java/mozilla/components/feature/prompts/dialog/MultiButtonDialogFragment.kt
@@ -4,18 +4,12 @@
 
 package mozilla.components.feature.prompts.dialog
 
-import android.annotation.SuppressLint
 import android.app.Dialog
 import android.content.DialogInterface
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.widget.CheckBox
 import androidx.appcompat.app.AlertDialog
-import mozilla.components.feature.prompts.R
 
-private const val KEY_MANY_ALERTS = "KEY_MANY_ALERTS"
 private const val KEY_SHOULD_DISMISS_ON_LOAD = "KEY_SHOULD_DISMISS_ON_LOAD"
-private const val KEY_USER_CHECK_BOX = "KEY_USER_CHECK_BOX"
 private const val KEY_POSITIVE_BUTTON_TITLE = "KEY_POSITIVE_BUTTON_TITLE"
 private const val KEY_NEGATIVE_BUTTON_TITLE = "KEY_NEGATIVE_BUTTON_TITLE"
 private const val KEY_NEUTRAL_BUTTON_TITLE = "KEY_NEUTRAL_BUTTON_TITLE"
@@ -24,12 +18,7 @@ private const val KEY_NEUTRAL_BUTTON_TITLE = "KEY_NEUTRAL_BUTTON_TITLE"
  * [android.support.v4.app.DialogFragment] implementation to display a confirm dialog,
  *  it can have up to three buttons, they could be positive, negative or neutral.
  */
-internal class MultiButtonDialogFragment : PromptDialogFragment() {
-
-    /**
-     * Tells if a checkbox should be shown for preventing this [sessionId] from showing more dialogs.
-     */
-    internal val hasShownManyDialogs: Boolean by lazy { safeArguments.getBoolean(KEY_MANY_ALERTS) }
+internal class MultiButtonDialogFragment : AbstractPromptTextDialogFragment() {
 
     internal val positiveButtonTitle: String? by lazy { safeArguments.getString(KEY_POSITIVE_BUTTON_TITLE) }
 
@@ -39,42 +28,18 @@ internal class MultiButtonDialogFragment : PromptDialogFragment() {
 
     override fun shouldDismissOnLoad() = safeArguments.getBoolean(KEY_SHOULD_DISMISS_ON_LOAD, true)
 
-    /**
-     * Stores the user's decision from the checkbox
-     * for preventing this [sessionId] from showing more dialogs.
-     */
-    private var userSelectionNoMoreDialogs: Boolean
-        get() = safeArguments.getBoolean(KEY_USER_CHECK_BOX)
-        set(value) {
-            safeArguments.putBoolean(KEY_USER_CHECK_BOX, value)
-        }
-
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         val builder = AlertDialog.Builder(requireContext())
             .setTitle(title)
             .setCancelable(true)
-            .setMessage(message)
             .setupButtons()
-        return (if (hasShownManyDialogs) addCheckbox(builder) else builder)
+        return setCustomMessageView(builder)
             .create()
     }
 
     override fun onCancel(dialog: DialogInterface) {
         super.onCancel(dialog)
         feature?.onCancel(sessionId)
-    }
-
-    @SuppressLint("InflateParams")
-    private fun addCheckbox(builder: AlertDialog.Builder): AlertDialog.Builder {
-        val inflater = LayoutInflater.from(requireContext())
-        val view = inflater.inflate(R.layout.mozac_feature_many_dialogs_checkbox_dialogs, null)
-        val checkBox = view.findViewById<CheckBox>(R.id.no_more_dialogs_check_box)
-        checkBox.setOnCheckedChangeListener { _, isChecked ->
-            userSelectionNoMoreDialogs = isChecked
-        }
-        builder.setView(view)
-
-        return builder
     }
 
     private fun AlertDialog.Builder.setupButtons(): AlertDialog.Builder {